Unfortunetally I have been a little lazy so there are a lot of changes.
The bigest cahnge to the ABP*.* scripts are that now they look for a file called coverage.json.
If they dont find this file then it only runs through the hetprofile so you have the info you need to create a het profile

Also made a lot of changes to bedForABP.py. These changes were made to support a pretty detailed track hub. It makes an html file, a copy of the psvgraph with the collapse name, and a big bed file now. These can then later be copied as a group into ~/public-html/track/ABP/

Started working on ProcessedCollapseAssembly.py which is what will allow me to create new genome wide things. For example CHM13 and yoruban.