ABP1 and ABP2 are now slightly modified to work off a config file called coverage.json this must include mincov maxcov mintotal and a project name
Fixed a bug in LocalAssembly.py

There were several bugs in marks original code the identify high coverage regions
So processcollapsesassembly.py is my version of it. I believe this is now at a point
where it works. It works on CHM1 for sure, still testing on CHM13.

bedForABP.y is a file that writes a bigbed file that can be put inot the genome browser.
it has some useful things, like a pic of the psv graph.