removed --fofn as input for SDA denovo and replaced with --input --input takes either a single pre aligned bam or an fofn of read files to align to the reference ont fofn files now must be fasta/fastq/fa/fq with an optional gz (minimap2 allowed formats)
if species for repeatmasker is a file it is passed to RM with -lib now.

the size of the reference is now read and used to addjust -I with minimap2.

added a line for sbatch to try and improve cluster submission.

TODO replace pre configured mem and thread areguments with a cluster.config.json.

TODO consider upgrading genomic concencous for pacbio and samtools to v1.10